feat: Split dino graphql object declarations
