+++ packages/server/src/dino/graphql/index.ts
@@ -1,0 +1,11 @@
+//
+//  index.ts
+//  dino-league
+//
+//  Created by d-exclaimation on 21 Dec 2022
+//
+
+export * from "./abstracts";
+export * from "./inputs";
+export * from "./others";
+export * from "./type";
